pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "artimuse"
version = "0.1.0"
description = "ArtiMuse: Aesthetic Assessment tools and demos"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "ArtiMuse Team"}]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

[project.optional-dependencies]
gradio = ["gradio", "matplotlib", "numpy", "Pillow"]
api = ["fastapi", "uvicorn"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["artimuse*"]

